+++ test/test_all.dart
@@ -1,6 +1,6 @@
 import 'package:test/test.dart';
 
-import 'api_test.dart' as api_generator;
+import 'api_generation_test.dart' as api_generator;
 
 void main() {
   group('api', api_generator.main);
